single rate limit package as source of truth
