+++ apps/api/src/index.ts
@@ -5,7 +5,7 @@ import { etag } from 'hono/etag'
 import { secureHeaders } from 'hono/secure-headers'
 import { sql } from '@workspace/db'
 import { buildContext } from './lib/context.ts'
-import { handleRateLimitError } from './lib/rate-limit.ts'
+import { handleRateLimitError } from '@workspace/rate-limit'
 import { requireSameOrigin, sessionMiddleware, type HonoEnv } from './middleware/session.ts'
 import { meRoute } from './routes/me.ts'
 import { usersRoute } from './routes/users.ts'
