+++ apps/api/src/lib/context.ts
@@ -8,7 +8,7 @@ import { loadEnv, type Env } from './env.ts'
 import { createCache, type Cache } from './cache.ts'
 import { createPubSub, type PubSub } from './pubsub.ts'
 import { createLogger, type Logger } from './logger.ts'
-import { makeRateLimit } from './rate-limit.ts'
+import { makeRateLimit } from '@workspace/rate-limit'
 
 export interface AppContext {
   env: Env
