+++ packages/rate-limit/package.json
@@ -7,10 +7,13 @@
     ".": "./src/index.ts"
   },
   "scripts": {
-    "typecheck": "tsc --noEmit"
+    "typecheck": "tsc --noEmit",
+    "smoke": "bun scripts/smoke.ts"
   },
   "dependencies": {
-    "ioredis": "^5.8.0"
+    "hono": "^4.7.11",
+    "ioredis": "^5.8.0",
+    "pino": "^9.6.0"
   },
   "devDependencies": {
     "@types/node": "^25.1.0",
